adicao de log, ao deletar, alterar

+++ src/main/java/com/fogo_na_panela_ws/fogo_na_panela_ws/dto/ProdutoDTO.java
@@ -3,14 +3,16 @@ package com.fogo_na_panela_ws.fogo_na_panela_ws.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ProdutoDTO {
     private Long id;
     private String nome;
     private String categoria;
-    private Double precoCompra;
-    private Double precoVenda;
+    private BigDecimal precoCompra;
+    private BigDecimal precoVenda;
     private Integer estoque;
     private String adicionado;
     private String ultimaAlteracao;
