+++ src/main/java/com/fogo_na_panela_ws/fogo_na_panela_ws/dto/ProdutoResponseDTO.java
@@ -3,14 +3,16 @@ package com.fogo_na_panela_ws.fogo_na_panela_ws.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ProdutoResponseDTO {
     private Long id;
     private String nome;
     private String categoria;
-    private Double precoCompra;
-    private Double precoVenda;
+    private BigDecimal precoCompra;
+    private BigDecimal precoVenda;
     private Integer estoque;
     private String adicionado;
     private String ultimaAlteracao;
